refactor. frontend apisix fetching setup done
