+++ frontend/src/App.tsx
@@ -1,10 +1,9 @@
 import { Button, Container, Heading } from "@chakra-ui/react";
 import Navbar from "./components/Navbar";
-import { usePosts } from "./api/hooks/usePosts";
+import { usePosts } from "./api/domain/posts/usePosts";
 
 function App() {
 	const { posts, page, loadPosts } = usePosts(1,3);
-	console.log(posts)
   return (
     <>
       <Navbar />
@@ -14,12 +13,17 @@ function App() {
         <Button color="brand.500">Click here lmao xd</Button>
       </Container>
 
+	{/* just some debug test fetching below - will be deleted */}
       <div>
+		<br></br>
+		<div>This below will only show posts if you have posts in your post db</div>
+		<div>Try Press "Prev" and "Next"</div>
+		<br></br>
         <button onClick={() => loadPosts(page - 1)}>Prev</button>
         <button onClick={() => loadPosts(page + 1)}>Next</button>
-
+		
         {posts.map((p) => (
-          <div key={p.id}>{p.title}</div>
+          <div key={p.u_id}>{p.title}</div>
         ))}
       </div>
     </>
