+++ frontend/src/api/domain/posts/post.ts
@@ -1,5 +1,5 @@
 export type Post = {
-  id: string;
+  u_id: string;
   title: string;
   content: string;
   author_id: string;
